# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CalendarSync OS API
  description: 複数のGoogleカレンダーを双方向同期するシステムのAPI仕様
  version: 1.0.0
  contact:
    name: CalendarSync OS Support
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.calendarsync.example.com/api
    description: 本番環境
...
paths:
  /auth/google:
    get:
      tags:
        - Auth
      summary: Google OAuth認証URL取得
      description: Google OAuth認証のURLを取得してリダイレクト
      responses:
        '302':
          description: Google認証ページへリダイレクト
        '500':
          description: エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/google/callback:
    get:
      tags:
        - Auth
      summary: OAuthコールバック処理
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: フロントエンドへリダイレクト
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Auth
      summary: 現在の認証状態を取得
      security:
        - sessionAuth: []
      responses:
        '200':
          description: アカウント情報
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: 未認証
  /auth/logout:
    post:
      tags:
        - Auth
      summary: ログアウト
      security:
        - sessionAuth: []
      responses:
        '200':
          description: ログアウト成功
  /accounts:
    get:
      tags:
        - Accounts
      summary: アカウント一覧取得
      security:
        - sessionAuth: []
      responses:
        '200':
          description: アカウント一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'

    delete:
      tags:
        - Accounts
      summary: アカウント削除
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 削除成功
        '404':
          description: アカウントが見つかりません
  /calendars:
    get:
      tags:
        - Calendars
      summary: カレンダー一覧取得
      security:
        - sessionAuth: []
      responses:
        '200':
          description: カレンダー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calendar'
  /calendars/{accountId}/sync:
    post:
      tags:
        - Calendars
      summary: カレンダー同期
      security:
        - sessionAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 同期成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  calendars:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calendar'
  /calendars/{id}:
    patch:
      tags:
        - Calendars
      summary: カレンダー設定更新
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sync_enabled:
                  type: boolean
                privacy_mode:
                  type: string
                  enum: [detail, busy-only]
                sync_direction:
                  type: string
                  enum: [bidirectional, readonly, writeonly]
      responses:
        '200':
          description: 更新成功
  /calendars/{calendarId}/events:
    post:
      tags:
        - Calendars
      summary: イベント作成
      security:
        - sessionAuth: []
      parameters:
        - name: calendarId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start_at
                - end_at
              properties:
                title:
                  type: string
                start_at:
                  type: string
                  format: date-time
                end_at:
                  type: string
                  format: date-time
                location:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: イベント作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: セッション認証
  schemas:
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        provider:
          type: string
          default: google
        workspace_flag:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Calendar:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        account_email:
          type: string
          format: email
        gcal_calendar_id:
          type: string
        name:
          type: string
          nullable: true
        sync_enabled:
          type: boolean
        privacy_mode:
          type: string
          enum: [detail, busy-only]
        sync_direction:
          type: string
          enum: [bidirectional, readonly, writeonly]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        location:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
          nullable: true